// frontend/src/lib/api/generated/models/DbStatsOutcomeStats.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DbStatsOutcomeStats = {
  commits: number;
  files_changed: number;
  loc_added: number;
  loc_removed: number;
  prs_merged?: number;
  prs_opened?: number;
  repos_active: number;
};